pyproject.toml
[tool.poetry]
name = "website"
version = "0.1.0"
description = ""
authors = ["tylercross <tyler.cross@me.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
python-fasthtml = "^0.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
